backend/djangoApp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
easyocr==1.7.2
eventlet==0.40.2
filelock==3.18.0
Flask==3.1.1
flask-sock==0.7.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
http_ece==1.2.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
mysqlclient==2.2.7
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
oauthlib==3.3.1
onnx==1.17.0
onnxruntime==1.22.1
onnxslim==0.1.61
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
py-vapid==1.9.2
pyclipper==1.3.0.post6
pycparser==2.22
PyJWT==2.10.1
pymongo==4.13.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-engineio==4.12.2
python-socketio==5.13.0
python3-openid==3.2.0
pytz==2025.2
pywebpush==2.0.3
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
scikit-image==0.25.2
scipy==1.16.0
shapely==2.1.1
simple-websocket==1.1.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zope.event==5.1.1
zope.interface==7.2
